+++ apps/api/v2/src/ee/event-types/inputs/enums/frequency.ts
@@ -1,0 +1,9 @@
+export enum Frequency {
+  YEARLY = 0,
+  MONTHLY = 1,
+  WEEKLY = 2,
+  DAILY = 3,
+  HOURLY = 4,
+  MINUTELY = 5,
+  SECONDLY = 6,
+}
