+++ apps/api/v2/src/ee/event-types/inputs/enums/period-type.ts
@@ -1,0 +1,5 @@
+export enum PeriodType {
+  UNLIMITED = "UNLIMITED",
+  ROLLING = "ROLLING",
+  RANGE = "RANGE",
+}
