+++ apps/api/v2/src/ee/event-types/inputs/enums/scheduling-type.ts
@@ -1,0 +1,5 @@
+export enum SchedulingType {
+  ROUND_ROBIN = "ROUND_ROBIN",
+  COLLECTIVE = "COLLECTIVE",
+  MANAGED = "MANAGED",
+}
